+++ src/Environment/Implementation.php
@@ -1,0 +1,68 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\CLI\Environment;
+
+use Innmind\Url\Path;
+use Innmind\Immutable\{
+    Map,
+    Sequence,
+    Str,
+    Attempt,
+    Maybe,
+};
+
+/**
+ * @psalm-immutable
+ * @internal
+ */
+interface Implementation
+{
+    /**
+     * True if the environment running the script is an interactive terminal
+     */
+    public function interactive(): bool;
+
+    /**
+     * @param ?positive-int $length
+     *
+     * @return array{Attempt<Str>, self}
+     */
+    public function read(?int $length = null): array;
+
+    /**
+     * @return Attempt<self>
+     */
+    public function output(Str $data): Attempt;
+
+    /**
+     * @return Attempt<self>
+     */
+    public function error(Str $data): Attempt;
+
+    /**
+     * @return Sequence<string>
+     */
+    public function arguments(): Sequence;
+
+    /**
+     * @return Map<string, string>
+     */
+    public function variables(): Map;
+
+    /**
+     * @param int<0, 254> $code
+     */
+    public function exit(int $code): self;
+
+    /**
+     * @return Maybe<ExitCode>
+     */
+    public function exitCode(): Maybe;
+    public function workingDirectory(): Path;
+
+    /**
+     * @return Sequence<array{Str, 'output'|'error'}>
+     */
+    public function outputted(): Sequence;
+}